PERF: n-loops

+++ tests/example_loop.py
@@ -39,6 +39,7 @@ gr=0
 subd=0
 otherside=True
 mydf = pd.read_csv( "CNAsynNeg_and_PDSporadicAsynNeg_ids_PPMI_only.csv" )
+ct=0
 for fn in list(mydf['filename']):
     path = construct_path(fn)
     if path is not None:
@@ -47,34 +48,39 @@ for fn in list(mydf['filename']):
         ccfn = [
             re.sub( ".nii.gz", "_"+ctype+"Rkappa.nii.gz" , path ), 
             re.sub( ".nii.gz", "_"+ctype+"R.nii.gz" , path ),
-            re.sub( ".nii.gz", "_"+ctype+"Rkappa.csv" , path ) ]
+            re.sub( ".nii.gz", "_"+ctype+"Rkappa.csv" , path ),
+            re.sub( ".nii.gz", "_"+ctype+"Rkappa.png" , path ) ]
         pcaud=[3,4]
         plabs=[4]
         if ctype == 'cit':
             mytl=18
-        xx = curvanato.t1w_caudcurv( cit, target_label=mytl, ventricle_label=vlab, 
-            prior_labels=pcaud, prior_target_label=plabs, subdivide=subd, grid=gr,
-            priorparcellation=tcaudR,  plot=False,
-            verbose=True )
-        ass
-        for j in range(2):
-            ants.image_write( xx[j], ccfn[j] )
-        xx[2].to_csv( ccfn[2] )
+        if not os.path.isfile(ccfn[3]):
+            xx = curvanato.t1w_caudcurv( cit, target_label=mytl, ventricle_label=vlab, 
+                prior_labels=pcaud, prior_target_label=plabs, subdivide=subd, grid=gr,
+                priorparcellation=tcaudR,  plot=False,
+                verbose=True )
+            ants.plot( xx[0], xx[1], crop=True, axis=2, nslices=21, ncol=7, filename=ccfn[3] )
+            for j in range(2):
+                ants.image_write( xx[j], ccfn[j] )
+            xx[2].to_csv( ccfn[2] )
         ######
         mytl=2
         ccfn = [
             re.sub( ".nii.gz", "_"+ctype+"Lkappa.nii.gz" , path ), 
             re.sub( ".nii.gz", "_"+ctype+"L.nii.gz" , path ),
-            re.sub( ".nii.gz", "_"+ctype+"Lkappa.csv" , path ) ]
+            re.sub( ".nii.gz", "_"+ctype+"Lkappa.csv" , path ),
+            re.sub( ".nii.gz", "_"+ctype+"Lkappa.png" , path ) ]
         print("Begin " + fn + " caud kap")
         pcaud=[1,2]
         plabs=[2]
-        xx = curvanato.t1w_caudcurv( cit, target_label=2, ventricle_label=vlab, 
-            prior_labels=pcaud, prior_target_label=plabs, subdivide=subd, grid=gr,
-            priorparcellation=tcaudL,  plot=True,
-            verbose=True )
-        for j in range(2):
-            ants.image_write( xx[j], ccfn[j] )
-        xx[2].to_csv( ccfn[2] )
+        if not os.path.isfile(ccfn[3]):
+            xx = curvanato.t1w_caudcurv( cit, target_label=2, ventricle_label=vlab, 
+                prior_labels=pcaud, prior_target_label=plabs, subdivide=subd, grid=gr,
+                priorparcellation=tcaudL,  plot=True,
+                verbose=True )
+            ants.plot( xx[0], xx[1], crop=True, axis=2, nslices=21, ncol=7, filename=ccfn[3] )
+            for j in range(2):
+                ants.image_write( xx[j], ccfn[j] )
+            xx[2].to_csv( ccfn[2] )
